move file io functions to main.js
