+++ src/spells.html
@@ -6,7 +6,7 @@
     <link rel="stylesheet" href="https://bootswatch.com/4/pulse/bootstrap.min.css">
     <link rel="stylesheet" type="text/css" href="../css/sidebar.css">
     <script type="text/javascript" src="http://ajax.googleapis.com/ajax/libs/jquery/1.6.2/jquery.min.js"></script>
-    <script type="text/javascript" src="./index"></script>
+    <script type="text/javascript" src="./main"></script>
 </head>
 
 <body>
@@ -60,8 +60,7 @@
     <script>
         //var jsonObj = { spells: [] }
         var table = document.getElementById("list")
-
-        //printFile()
+        printFile()
 
         $.getJSON('../charsheet.json', function (data) {
             $.each(data.spells, function (i, f) {
